numerous bug fixes to File-related classes.
git-svn-id: https://svn.code.sf.net/p/quorum/code/trunk@580 df4a1cb3-768c-44ce-9dd6-df0ee0bb247f

+++ plugins/FilePlugin/src/plugins/quorum/Libraries/System/FileRandomAccess.java
@@ -38,7 +38,7 @@ public class FileRandomAccess {
         }
     }
 
-    public long GetPosition() throws InputOutputError {
+    public double GetPosition() throws InputOutputError {
         try {
             return inst.GetPosition();
         } catch (IOException ex) {
@@ -49,9 +49,9 @@ public class FileRandomAccess {
         }
     }
     
-    public void SetPositionNative(long position) throws InputOutputError {
+    public void SetPositionNative(double position) throws InputOutputError {
         try {
-            inst.SetPositionNative(position);
+            inst.SetPositionNative((long)position);
         } catch (IOException ex) {
             InputOutputError e = new InputOutputError();
             if (ex.getMessage() != null)
